combining two modals
